% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_modelfit_mlm.R
\name{allometree_mlm}
\alias{allometree_mlm}
\title{Fit mixed-effects or multilevel models to all species}
\usage{
allometree_mlm(
  data,
  species = "species",
  response = "height",
  predictor = "diameter"
)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree.}

\item{species}{Column name of the species variable in both the dataframes
\code{data} and \code{ref_table}. Defaults to \code{species}.}

\item{response}{Column name of the response variable.
Defaults to \code{height}.}

\item{predictor}{Column name of the predictor variable.
Defaults to \code{diameter}.}
}
\value{
A list of 4 elements:
\describe{
\item{models_rank}{A model selection table of all types of mixed-effects models}
\item{best_model}{The best model object.}
\item{R2}{The conditional and marginal pseudo-\eqn{R^2} of the best model.}
\item{warnings}{Warning messages, if any, spit from the models.
These usually indicate failure of model convergence.}
}
}
\description{
Fit mixed-effects or multilevel models to the dataset pooled across all species, using the \code{lme4::lmer} function under the hood.
}
\examples{
m_test <- allometree_mlm(data = subset(urbantrees, height < 40))
m_test
}
