% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_newX.R
\name{generate_newX}
\alias{generate_newX}
\title{Helper function to generate new predictors for each species for simulation/prediction}
\usage{
generate_newX(
  data,
  response = "height",
  predictor = "diameter",
  species = "species",
  length.out = 100
)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree of a particular species.}

\item{response}{Column name of the response variable in \code{data}. Defaults to
\code{height}.}

\item{predictor}{Column name of the predictor variable in \code{data}. Defaults to
\code{diameter}.}

\item{species}{Column name of the species variable in \code{data}. Defaults to \code{species}.}

\item{length.out}{Number of new X values to generate for each species. Defaults to 100. Set a higher value for greater resolution at the cost of computational time.}
}
\value{
A new dataset containing the columns species, new X, and whether X is extrapolated for each species.
}
\description{
Helper function to generate new predictors for each species for simulation/prediction
}
\examples{
newdat <- generate_newX(urbantrees)
newdat
}
