% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_x.R
\name{generate_x}
\alias{generate_x}
\title{Generate new predictor values for each species in a dataset}
\usage{
generate_x(
  data,
  response = "height",
  predictor = "diameter",
  species = "species",
  extrapolate = NULL,
  length.out = 100
)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree of a particular species.}

\item{response}{Column name of the response variable in \code{data}. Defaults to
\code{height}.}

\item{predictor}{Column name of the predictor variable in \code{data}. Defaults to
\code{diameter}.}

\item{species}{Column name of the species variable in \code{data}. Defaults to \code{species}.}

\item{extrapolate}{Numeric vector of 2 elements (e.g. \code{c(0,4)}), representing
the upper and lower bounds of extrapolation. Defaults to \code{NULL} for no
extrapolation.}

\item{length.out}{Number of new values to generate for each species. Defaults to 100. Set a higher value for greater resolution at the cost of computational time.}
}
\value{
A dataframe with columns:
\describe{
\item{species}{Name of tree species.}
\item{predictor}{Newly generated predictor values.}
\item{extrapolated}{Indicates whether the predictor (x) values are extrapolated beyond the supplied dataset.
Either 'High', 'Low', or 'No' (not extrapolated).}
}
}
\description{
Helper function to generate a sequence of new values for each species,
based on the range of values for the predictor variable in the \code{data}.
Upward and downward extrapolation to specified values can be made.
Predictions can subsequently be made on the generated values.
}
\examples{
newdata <- generate_x(urbantrees,
                      extrapolate = c(0,4),
                      response = "height", predictor = "diameter")
head(newdata)

}
