% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_modelselect.R
\name{mix_modelselect}
\alias{mix_modelselect}
\title{Fit linear mixed-effects model across all species}
\usage{
mix_modelselect(
  data,
  species = "species",
  response = "height",
  predictor = "diameter"
)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree.}

\item{species}{Column name of the species variable in \code{data}. Defaults to \code{species}.}

\item{response}{Column name of the response variable.
Defaults to \code{height}.}

\item{predictor}{Column name of the predictor variable.
Defaults to \code{diameter}.}
}
\value{
A list of 5 elements:
\describe{
\item{models_rank}{A model selection table of all the types of mixed-effects models considered,
ranked in order of ascending Aikake's Information Criterion corrected for small sample sizes (AICc).
Model details can be found in \code{?eqns_info} and \code{data(eqns_info)}.}
\item{best_model}{The best-fit model object.}
\item{R2}{The conditional and marginal pseudo-\eqn{R^2} of the best-fit model.}
\item{CF}{Correction factor used to adjust predicted values if response variable is transformed
(incorporated into reported parameters).}
\item{warnings}{Warning messages, if any, spit from the models.
These usually indicate failure of model convergence.}
}
}
\description{
Fit data to linear mixed-effects model with 'species' specified as the random effect,
using the \code{lme4::lmer} function under the hood. The full list of allometric equations that are
considered can be found in \code{?eqns_info} and \code{data(eqns_info)}.
}
\examples{
data(urbantrees)

\dontrun{
mix_modelselect(data = urbantrees,
                species = "species",
                response = "height", predictor = "diameter")
}

}
\seealso{
Other mixed-effects model functions: 
\code{\link{mix_simulate}()}
}
\concept{mixed-effects model functions}
