% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_simulate.R
\name{mix_simulate}
\alias{mix_simulate}
\title{Make predictions on simulated data using mixed-effects model}
\usage{
mix_simulate(
  data,
  modelselect,
  level = 0.95,
  extrapolate = NULL,
  length.out = 100,
  stat = "median",
  n.sims = 1000,
  response = "height",
  predictor = "diameter",
  species = "species",
  ...
)
}
\arguments{
\item{data}{Dataframe used to generate data and their predictions the using mixed-effects model.
Columns should contain the species and variables of interest. Each row is a measurement for an individual tree.}

\item{modelselect}{Output from the \code{\link[=mix_modelselect]{mix_modelselect()}} function.}

\item{level}{Level of confidence for the prediction interval. Defaults to
\code{0.95}.}

\item{extrapolate}{Numeric vector of 2 elements (e.g. \code{c(0,4)}), representing
the upper and lower bounds of extrapolation. Defaults to \code{NULL} for no
extrapolation.}

\item{length.out}{Number of new X values to generate for each species. Defaults to 100. Set a higher value for greater resolution at the cost of computational time.}

\item{stat}{Specify whether the \code{"median"} or \code{"mean"} of simulated intervals are used.}

\item{n.sims}{Number of bootstrapped simulations to generate the prediction intervals. Defaults to \code{1000}.}

\item{response}{Column name of the response variable in \code{data}. Defaults to
\code{height}.}

\item{predictor}{Column name of the predictor variable in \code{data}. Defaults to
\code{diameter}.}

\item{species}{Column name of the species variable in \code{data}. Defaults to \code{species}.}

\item{...}{Additional arguments passed to \code{\link[merTools:predictInterval]{merTools::predictInterval()}}}
}
\value{
Dataframe containing data generated from \code{\link[=generate_x]{generate_x()}}, and their predictions (with intervals).
}
\description{
Data is simulated for each species (if multiple species are present) based on their respective ranges of the
predictor variable, which can be extrapolated to values defined by the user. The mixed-effects model is used
to predict values for the response variable, as well as it's prediction interval.
Necessary bias-corrections are made if the mixed-effects model has a transformed response variable.
}
\examples{
data(urbantrees)

\dontrun{
model <- mix_modelselect(urbantrees)
results <- mix_simulate(data = urbantrees, modelselect = model)
head(results)
}

}
\seealso{
\code{\link[=mix_modelselect]{mix_modelselect()}}, \code{\link[=generate_x]{generate_x()}}, \code{\link[merTools:predictInterval]{merTools::predictInterval()}}
}
