% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_modelfit.R
\name{sp_modelfit_multi}
\alias{sp_modelfit_multi}
\title{Fit data to pre-selected models across multiple species}
\usage{
sp_modelfit_multi(
  data,
  ref_table,
  species = "species",
  modelcode = "modelcode",
  response = "height",
  predictor = "diameter"
)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree of a particular species.}

\item{ref_table}{Dataframe containing an allometric equation for each tree
species, in the form of a model code. Each row is a unique species.}

\item{species}{Column name of the species variable in both the dataframes
\code{data} and \code{ref_table}. Defaults to \code{species}.}

\item{modelcode}{Column name containing the model codes in \code{ref_table}. Refer
to \code{data(eqns_info)} for more information on model codes.}

\item{response}{Column name of the response variable in \code{data}. Defaults to
\code{height}.}

\item{predictor}{Column name of the predictor variable in \code{data}. Defaults to
\code{diameter}.}
}
\value{
A list of 2 elements:
\enumerate{
\item List of each species' resulting model object, named \code{sp_models}.
\item Table showing each species' resulting model information, named
\code{sp_models_info}.
}
\subsection{sp_models_info}{

\code{sp_models_info} is a dataframe with the following variables:
\describe{
\item{species}{Name of tree species}
\item{modelcode}{Model code for the allometric equation used}
\item{a, b, c, d, e}{Parameter estimates}
\item{response_geom_mean}{Geometric mean of the response variable used in
calculation of AICc (only for transformed models)}
\item{correctn_factor}{Bias correction factor to use on model predictions
(only for transformed models)}
\item{predictor_min, predictor_max}{Range of the predictor variable within
the data used to generate the model (i.e. threshold values for model extrapolation)}
\item{residual_SE}{Residual standard error of the model}
\item{mean_SE}{Mean standard error of the model}
\item{adj_R2}{Adjusted \eqn{R^2} of the model}
\item{n}{Sample size (no. of trees used to fit model)}
}
}
}
\description{
Wrapper to run \code{sp_modelfit()} across multiple species. Data is fit to
pre-selected allometric equations for each species, as defined within
\code{ref_table}. The full list of allometric equations that may be considered in
\code{ref_table} can be found in \code{?eqns_info} and \code{data(eqns_info)}.
}
\examples{
# first select best-fit model for all species in data
data(urbantrees)
selected <- sp_modelselect_multi(urbantrees, species = 'species', response = 'height', predictor = 'diameter')

# use function
results <- sp_modelfit_multi(
  urbantrees, # any dataset with similar species can be used (in this case, we re-use the same data)
  ref_table = selected$sp_models_info, # refer to reference table from results above
  species = 'species', modelcode = 'modelcode',
  response = 'height', predictor = 'diameter'
)

results$sp_models[[1]] # model object for first species in list

results$sp_models_info # summary of fitted models

}
\seealso{
\code{\link[=sp_modelfit]{sp_modelfit()}} to fit pre-selected model for one species.

\code{\link[=sp_modelselect]{sp_modelselect()}} to select a best-fit model for one species.

\code{\link[=sp_modelselect_multi]{sp_modelselect_multi()}} to select best-fit models across multiple species.

Other single-species model functions: 
\code{\link{sp_modelfit}()},
\code{\link{sp_modelselect_multi}()},
\code{\link{sp_modelselect}()}
}
\concept{single-species model functions}
