% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_modelselect.R
\name{sp_modelselect}
\alias{sp_modelselect}
\title{Select a best-fit linear model for one species}
\usage{
sp_modelselect(data, response = "height", predictor = "diameter")
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree.}

\item{response}{Column name of the response variable. Defaults to \code{height}.}

\item{predictor}{Column name of the predictor variable. Defaults to
\code{diameter}.}
}
\value{
A list of 3 elements: \describe{ \item{all_models_rank}{Table showing
models ranked by AICc value.} \item{best_model}{Best-fit model object.}
\item{best_model_info}{Table showing information on the best-fit model.} }
\subsection{best_model_info}{

A dataframe with the following variables: \describe{ \item{modelcode}{Model
code for the best-fit equation.} \item{a, b, c, d, e}{Parameter estimates.}
\item{response_geom_mean}{Geometric mean of the response variable used in
calculation of AICc (only for transformed models).}
\item{correctn_factor}{Bias correction factor to use on model predictions
(only for transformed models).} \item{predictor_min, predictor_max}{Range of
the predictor variable within the data used to generate the model.}
\item{response_min, response_max}{Range of the response variable within the
data used to generate the model.} \item{residual_SE}{Residual standard error
of the model.} \item{mean_SE}{Mean standard error of the model.}
\item{adj_R2}{Adjusted \eqn{R^2} of the model.} \item{n}{Sample size (no. of
trees used to fit model).} }
}
}
\description{
Select a best-fit equation for one species, based on the lowest
bias-corrected Aikaike’s information criterion (AICc).
}
\details{
All allometric equations considered (and ranked) can be found in \code{?eqns_info}
and \code{data(eqns_info)}. To make the AICc values of equations with a
transformed response variable comparable to untransformed equations,
\eqn{log(y_{i})} is multiplied by the geometric mean of the response variable
in \code{data}.
}
\examples{
data(urbantrees)
Alb_sam <- urbantrees[urbantrees$species == 'Albizia saman', ]  # subset data for 1 species
results <- sp_modelselect(Alb_sam, response = 'height', predictor = 'diameter')

head(results$all_models_rank)

results$best_model

results$best_model_info

}
\references{
McPherson E. G., van Doorn N. S. & Peper P. J. (2016) Urban Tree
Database and Allometric Equations. \emph{General Technical Report PSW-GTR-253,
USDA Forest Service}, 86.

Xiao, X., White, E. P., Hooten, M. B., & Durham, S. L. (2011). On the use
of log-transformation vs. nonlinear regression for analyzing biological
power laws. \emph{Ecology}, \strong{92}(10), 1887–1894.

Burnham, K. P., & Anderson, D. R. (2004). Multimodel inference:
Understanding AIC and BIC in model selection. \emph{Sociological Methods and
Research}, \strong{33}(2), 261–304.
}
\seealso{
\code{\link[=sp_modelselect_multi]{sp_modelselect_multi()}} to select best-fit models across multiple
species.

\code{\link[=sp_modelfit]{sp_modelfit()}} to fit a pre-selected model for one species.

\code{\link[=sp_modelfit_multi]{sp_modelfit_multi()}} to fit pre-selected models across multiple species.

Other single-species model functions: 
\code{\link{sp_modelfit_multi}()},
\code{\link{sp_modelfit}()},
\code{\link{sp_modelselect_multi}()},
\code{\link{sp_predict}()},
\code{\link{sp_simulate}()}
}
\concept{single-species model functions}
