% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_modelselect.R
\name{sp_modelselect_multi}
\alias{sp_modelselect_multi}
\title{Select best-fit linear models across multiple species}
\usage{
sp_modelselect_multi(
  data,
  species = "species",
  response = "height",
  predictor = "diameter"
)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a measurement
for an individual tree of a particular species.}

\item{species}{Column name of the species variable. Defaults to \code{species}.}

\item{response}{Column name of the response variable. Defaults to \code{height}.}

\item{predictor}{Column name of the predictor variable. Defaults to \code{diameter}.}
}
\value{
A list of 3 elements:
\enumerate{
\item List of tables showing each species' candidate models ranked by AICc value,
named \code{sp_models_rank}.
\item List of each species' best-fit model object, named \code{sp_models}.
\item Table showing each species' best-fit model information, named \code{sp_models_info}.
}
\subsection{sp_models_info}{

\code{sp_models_info} is a dataframe with the following variables:
\describe{
\item{species}{Name of tree species}
\item{modelcode}{Model code for the best-fit equation}
\item{a, b, c, d, e}{Parameter estimates}
\item{response_geom_mean}{Geometric mean of the response variable used in
calculation of AICc (only for transformed models)}
\item{correctn_factor}{Bias correction factor to use on model predictions
(only for transformed models)}
\item{predictor_min, predictor_max}{Range of the predictor variable within
the data used to generate the model (i.e. threshold values for model extrapolation)}
\item{residual_SE}{Residual standard error of the model}
\item{mean_SE}{Mean standard error of the model}
\item{adj_R2}{Adjusted \eqn{R^2} of the model}
\item{n}{Sample size (no. of trees used to fit model)}
}
}
}
\description{
Wrapper function that runs \code{sp_modelselect()} across multiple species in a dataframe.
A single best-fit equation is selected per species, based on the lowest AICc value.
All allometric equations considered (and ranked) can be found in \code{?eqns_info} and \code{data(eqns_info)}.
}
\examples{
data(urbantrees)
results <- sp_modelselect_multi(urbantrees, species = 'species',
                                response = 'height', predictor = 'diameter')

head(results$sp_models_rank[[1]]) # Highly-ranked models for first species in list

results$sp_models[[1]] # model object for first species in list

results$sp_models_info # summary of best-fit models

}
\seealso{
\code{\link[=sp_modelselect]{sp_modelselect()}} to select a best-fit model for one species.

\code{\link[=sp_modelfit]{sp_modelfit()}} to fit a pre-selected model for one species.

\code{\link[=sp_modelfit_multi]{sp_modelfit_multi()}} to fit pre-selected models across multiple species.

Other single-species model functions: 
\code{\link{sp_modelfit_multi}()},
\code{\link{sp_modelfit}()},
\code{\link{sp_modelselect}()},
\code{\link{sp_predict}()},
\code{\link{sp_simulate}()}
}
\concept{single-species model functions}
