% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_predict.R
\name{sp_predict}
\alias{sp_predict}
\title{Make predictions for all species in a dataset using single-species linear
models}
\usage{
sp_predict(
  data,
  models,
  ref_table,
  level = 0.95,
  species = "species",
  predictor = "diameter",
  cf = "correctn_factor",
  geom_mean = "response_geom_mean"
)
}
\arguments{
\item{data}{Dataframe with columns containing the species and variables of
interest. Species names in the \code{species} column should be present in those
within \code{ref_table}, as well as those in \code{models}.}

\item{models}{A named list of each species' linear regression models.
\code{names(models)} should correspond to species names in \code{data} and
\code{ref_table}.}

\item{ref_table}{Dataframe containing information to correct bias introduced
in models with a transformed response variable. It should include columns
for \code{species}, \code{cf}, and \code{geom_mean}.}

\item{level}{Level of confidence for the prediction interval. Defaults to
\code{0.95}.}

\item{species}{Column name of the species variable in \code{data} and \code{ref_table}.
Defaults to \code{species}.}

\item{predictor}{Column name of the predictor variable in \code{data}. Defaults to
\code{diameter}.}

\item{cf}{Column name of the bias correction factor in \code{ref_table}. Defaults
to \code{correctn_factor}.}

\item{geom_mean}{Column name of the geometric mean of response variable in
\code{ref_table}, that was used in to fit the \code{models}. Defaults to
\code{response_geom_mean}.}
}
\value{
Dataframe of input \code{data} with columns appended: \describe{
\item{fit}{Predicted value for the response variable.} \item{lwr}{Lower
bound of the prediction interval, based on the input argument \code{level}.}
\item{upr}{Upper bound of the prediction interval, based on the input
argument \code{level}.} }
}
\description{
Make predictions across rows in a dataset that may contain multiple species.
The model associated with each species is used to predict values for the
response variable, as well as it's prediction interval. Necessary
bias-corrections are made for species with models that have a transformed
response variable.
}
\examples{
# first select best-fit model
data(urbantrees)
Alb_sam <- urbantrees[urbantrees$species == 'Albizia saman', ]  # we use one species as an example
results <- sp_modelselect_multi(Alb_sam,
                                species = 'species',
                                response = 'height',
                                predictor = 'diameter')

# simulate data
predict_range <- results$sp_models_info[ ,c('species','predictor_min', 'predictor_max')]
predict_range_full <- as.data.frame(apply(predict_range, 1,
                                          function(x) seq(x['predictor_min'], x['predictor_max'],
                                          length.out = 100)))
colnames(predict_range_full) <- predict_range$species
predict_range_full <- tidyr::pivot_longer(predict_range_full,
                                          cols = colnames(predict_range_full),
                                          names_to = 'species',
                                          values_to = 'predictor')

# run function
predictions <- sp_predict(predict_range_full,
                          models = results$sp_models,
                          ref_table = results$sp_models_info,
                          predictor = 'predictor')
head(predictions)

}
\seealso{
\code{\link[=sp_simulate]{sp_simulate()}} to run \code{sp_predict()} on simulated data.

Other single-species model functions: 
\code{\link{sp_modelfit_multi}()},
\code{\link{sp_modelfit}()},
\code{\link{sp_modelselect_multi}()},
\code{\link{sp_modelselect}()},
\code{\link{sp_simulate}()}
}
\concept{single-species model functions}
