% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_predict_mlm.R
\name{sp_predict_mlm}
\alias{sp_predict_mlm}
\title{Predictions or simulation for mixed-effects or multilevel models}
\usage{
sp_predict_mlm(data, models, level = 0.95, stat = "median", n.sims = 1000, ...)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree.}

\item{models}{Output from the \code{\link[=allometree_mlm]{allometree_mlm()}} function.}

\item{level}{The width of the prediction intervals}

\item{stat}{Take the median or mean of simulated intervals}

\item{n.sims}{Number of bootstrap simulation for the prediction intervals}

\item{...}{Additional arguments passed to \code{\link[merTools:predictInterval]{merTools::predictInterval()}}}
}
\value{
A dataframe containing generated newX from \code{\link[=generate_newX]{generate_newX()}} and predictions (with intervals)
}
\description{
Predictions or simulation for mixed-effects or multilevel models
}
\examples{
dat <- subset(urbantrees, height < 40)
m_test <- allometree_mlm(dat)
pred_mlm <- sp_predict_mlm(dat, m_test)
pred_mlm
}
\seealso{
\code{\link[=allometree_mlm]{allometree_mlm()}}, \code{\link[=generate_newX]{generate_newX()}}, \code{\link[merTools:predictInterval]{merTools::predictInterval()}}
}
