% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_modelfit.R
\name{ss_modelfit}
\alias{ss_modelfit}
\title{Fit data to a specified linear model for one species}
\usage{
ss_modelfit(data, modelcode, response = "height", predictor = "diameter")
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree.}

\item{modelcode}{Character string of the model code for the selected
allometric equation. Refer to \code{?eqns_info} and \code{data(eqns_info)} for more
information.}

\item{response}{Column name of the response variable. Defaults to \code{height}.}

\item{predictor}{Column name of the predictor variable. Defaults to
\code{diameter}.}
}
\value{
A list of 2 elements: \describe{ \item{fitted_model}{Resulting model
object.} \item{fitted_model_info}{Table showing information on the
resulting model.} }
\subsection{fitted_model_info}{

A dataframe with the following variables: \describe{ \item{modelcode}{Model
code for the allometric equation used.} \item{a, b, c, d, e}{Parameter
estimates.} \item{response_geom_mean}{Geometric mean of the response
variable used in calculation of AICc (only for transformed models).}
\item{correctn_factor}{Bias correction factor to use on model predictions
(only for transformed models).} \item{predictor_min, predictor_max}{Range of
the predictor variable within the data used to generate the model.}
\item{response_min, response_max}{Range of the response variable within the
data used to generate the model.} \item{residual_SE}{Residual standard error
of the model.} \item{mean_SE}{Mean standard error of the model.}
\item{adj_R2}{Adjusted \eqn{R^2} of the model.} \item{n}{Sample size (no. of
trees used to fit model).} }
}
}
\description{
Fit data to a specified allometric equation, for one species. Allometric
equations that may be considered as an input to this function can be found in
\code{?eqns_info} and \code{data(eqns_info)}.
}
\examples{
data(urbantrees)
Alb_sam <- urbantrees[urbantrees$species == 'Albizia saman', ]
results <- ss_modelfit(Alb_sam,
                       modelcode = 'quad_w1', # manually specify equation to use
                       response = 'height', predictor = 'diameter')

results$fitted_model

results$fitted_model_info

}
\references{
McPherson E. G., van Doorn N. S. & Peper P. J. (2016) Urban Tree
Database and Allometric Equations. \emph{General Technical Report PSW-GTR-253,
USDA Forest Service}, 86.
}
\seealso{
\code{\link[=ss_modelfit_multi]{ss_modelfit_multi()}} to fit specified models across multiple
species.

\code{\link[=ss_modelselect]{ss_modelselect()}} to select a best-fit model for one species.

\code{\link[=ss_modelselect_multi]{ss_modelselect_multi()}} to select best-fit models across multiple species.

Other single-species model functions: 
\code{\link{ss_modelfit_multi}()},
\code{\link{ss_modelselect_multi}()},
\code{\link{ss_modelselect}()},
\code{\link{ss_predict}()},
\code{\link{ss_simulate}()}
}
\concept{single-species model functions}
