% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_modelfit.R
\name{ss_modelfit_multi}
\alias{ss_modelfit_multi}
\title{Fit data to specified linear models across multiple species}
\usage{
ss_modelfit_multi(
  data,
  ref_table,
  species = "species",
  modelcode = "modelcode",
  response = "height",
  predictor = "diameter"
)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree of a particular species.}

\item{ref_table}{Dataframe containing an allometric equation for each tree
species, in the form of a model code. Each row is a unique species.}

\item{species}{Column name of the species variable in both the dataframes
\code{data} and \code{ref_table}. Defaults to \code{species}.}

\item{modelcode}{Column name containing the model codes in \code{ref_table}. Refer
to \code{data(eqns_info)} for more information on model codes.}

\item{response}{Column name of the response variable in \code{data}. Defaults to
\code{height}.}

\item{predictor}{Column name of the predictor variable in \code{data}. Defaults to
\code{diameter}.}
}
\value{
A list of 2 elements: \describe{ \item{ss_models}{List of each
species' resulting model object.} \item{ss_models_info}{Table showing each
species' resulting model information.} }
\subsection{ss_models_info}{

A dataframe with the following variables: \describe{ \item{species}{Name of
tree species.} \item{modelcode}{Model code for the allometric equation used.}
\item{a, b, c, d, e}{Parameter estimates.}
\item{response_geom_mean}{Geometric mean of the response variable used in
calculation of AICc (only for transformed models).}
\item{correctn_factor}{Bias correction factor to use on model predictions
(only for transformed models).} \item{predictor_min, predictor_max}{Range of
the predictor variable within the data used to generate the model.}
\item{response_min, response_max}{Range of the response variable within the
data used to generate the model.} \item{residual_SE}{Residual standard error
of the model.} \item{mean_SE}{Mean standard error of the model.}
\item{adj_R2}{Adjusted \eqn{R^2} of the model.} \item{n}{Sample size (no. of
trees used to fit model).} }
}
}
\description{
Wrapper function that runs \code{ss_modelfit()} across multiple species. Data is
fit to specified allometric equations for each species, as defined within
\code{ref_table}. The full list of allometric equations that may be considered in
\code{ref_table} can be found in \code{?eqns_info} and \code{data(eqns_info)}.
}
\examples{
# first select best-fit model for all species in data
data(urbantrees)
selected <- ss_modelselect_multi(urbantrees, species = 'species',
                                 response = 'height', predictor = 'diameter')

# use function
results <- ss_modelfit_multi(
  urbantrees, # any data with similar species (re-use same data in this case)
  ref_table = selected$ss_models_info,
  species = 'species', modelcode = 'modelcode',
  response = 'height', predictor = 'diameter'
)

results$ss_models[[1]] # model object for first species in list

results$ss_models_info # summary of fitted models

}
\seealso{
\code{\link[=ss_modelfit]{ss_modelfit()}} to fit a specified model for one species.

\code{\link[=ss_modelselect]{ss_modelselect()}} to select a best-fit model for one species.

\code{\link[=ss_modelselect_multi]{ss_modelselect_multi()}} to select best-fit models across multiple species.

Other single-species model functions: 
\code{\link{ss_modelfit}()},
\code{\link{ss_modelselect_multi}()},
\code{\link{ss_modelselect}()},
\code{\link{ss_predict}()},
\code{\link{ss_simulate}()}
}
\concept{single-species model functions}
