% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_modelselect.R
\name{ss_modelselect_multi}
\alias{ss_modelselect_multi}
\title{Select best-fit linear models across multiple species}
\usage{
ss_modelselect_multi(
  data,
  species = "species",
  response = "height",
  predictor = "diameter"
)
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree of a particular species.}

\item{species}{Column name of the species variable. Defaults to \code{species}.}

\item{response}{Column name of the response variable. Defaults to \code{height}.}

\item{predictor}{Column name of the predictor variable. Defaults to
\code{diameter}.}
}
\value{
A list of 3 elements:  \describe{ \item{ss_models_rank}{List of
tables showing each species' candidate models ranked by AICc value.}
\item{ss_models}{List of each species' best-fit model object.}
\item{ss_models_info}{Table showing each species' best-fit model
information.} }
\subsection{ss_models_info}{

A dataframe with the following variables: \describe{ \item{species}{Name of
tree species.} \item{modelcode}{Model code for the best-fit equation.}
\item{a, b, c, d, e}{Parameter estimates.}
\item{response_geom_mean}{Geometric mean of the response variable used in
calculation of AICc (only for transformed models).}
\item{correctn_factor}{Bias correction factor to use on model predictions
(only for transformed models).} \item{predictor_min, predictor_max}{Range of
the predictor variable within the data used to generate the model.}
\item{response_min, response_max}{Range of the response variable within the
data used to generate the model.} \item{residual_SE}{Residual standard error
of the model.} \item{mean_SE}{Mean standard error of the model.}
\item{adj_R2}{Adjusted \eqn{R^2} of the model.} \item{n}{Sample size (no. of
trees used to fit model).} }
}
}
\description{
Wrapper function that runs \code{ss_modelselect()} across multiple species in a
dataframe. A single best-fit equation is selected per species, based on the
lowest AICc value. All allometric equations considered (and ranked) can be
found in \code{?eqns_info} and \code{data(eqns_info)}.
}
\examples{
data(urbantrees)
results <- ss_modelselect_multi(urbantrees, species = 'species',
                                response = 'height', predictor = 'diameter')

head(results$ss_models_rank[[1]]) # Highly-ranked models for 1st species in list

results$ss_models[[1]] # model object for 1st species in list

results$ss_models_info # summary of best-fit models

}
\seealso{
\code{\link[=ss_modelselect]{ss_modelselect()}} to select a best-fit model for one species.

\code{\link[=ss_modelfit]{ss_modelfit()}} to fit a pre-selected model for one species.

\code{\link[=ss_modelfit_multi]{ss_modelfit_multi()}} to fit pre-selected models across multiple species.

Other single-species model functions: 
\code{\link{ss_modelfit_multi}()},
\code{\link{ss_modelfit}()},
\code{\link{ss_modelselect}()},
\code{\link{ss_predict}()},
\code{\link{ss_simulate}()}
}
\concept{single-species model functions}
