% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_predict.R
\name{ss_predict}
\alias{ss_predict}
\title{Make predictions for all species in a dataset using single-species linear
models}
\usage{
ss_predict(
  data,
  models,
  ref_table,
  level = 0.95,
  species = "species",
  predictor = "diameter",
  cf = "correctn_factor",
  geom_mean = "response_geom_mean"
)
}
\arguments{
\item{data}{Dataframe with columns containing the species and variables of
interest. Species names in the \code{species} column should be present in those
within \code{ref_table}, as well as those in \code{models}.}

\item{models}{A named list of each species' linear regression models.
\code{names(models)} should correspond to species names in \code{data} and
\code{ref_table}.}

\item{ref_table}{Dataframe containing information to correct bias introduced
in models with a transformed response variable. It should include columns
for \code{species}, \code{cf}, and \code{geom_mean}.}

\item{level}{Level of confidence for the prediction interval. Defaults to
\code{0.95}.}

\item{species}{Column name of the species variable in \code{data} and \code{ref_table}.
Defaults to \code{species}.}

\item{predictor}{Column name of the predictor variable in \code{data}. Defaults to
\code{diameter}.}

\item{cf}{Column name of the bias correction factor in \code{ref_table}. Defaults
to \code{correctn_factor}.}

\item{geom_mean}{Column name of the geometric mean of response variable in
\code{ref_table}, that was used in to fit the \code{models}. Defaults to
\code{response_geom_mean}.}
}
\value{
Dataframe of input \code{data} with columns appended: \describe{
\item{fit}{Predicted value for the response variable.} \item{lwr}{Lower
bound of the prediction interval, based on the input argument \code{level}.}
\item{upr}{Upper bound of the prediction interval, based on the input
argument \code{level}.} }
}
\description{
Make predictions across rows in a dataset that may contain multiple species.
The model associated with each species is used to predict values for the
response variable, as well as it's prediction interval. Necessary
bias-corrections are made for species with models that have a transformed
response variable.
}
\examples{
# first select best-fit model
data(urbantrees)
Alb_sam <- urbantrees[urbantrees$species == 'Albizia saman', ]  # we use one species as an example
results <- ss_modelselect_multi(Alb_sam,
                                response = 'height',
                                predictor = 'diameter')

# generate data for subsequent predictions
newdata <- generate_x(Alb_sam,
                      response = "height", predictor = "diameter")

# run function
predictions <- ss_predict(newdata,
                          models = results$ss_models,
                          ref_table = results$ss_models_info,
                          predictor = "predictor")
head(predictions)

}
\seealso{
\code{\link[=ss_simulate]{ss_simulate()}} to run \code{ss_predict()} on simulated data.

Other single-species model functions: 
\code{\link{ss_modelfit_multi}()},
\code{\link{ss_modelfit}()},
\code{\link{ss_modelselect_multi}()},
\code{\link{ss_modelselect}()},
\code{\link{ss_simulate}()}
}
\concept{single-species model functions}
