% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-urbantrees.R
\docType{data}
\name{urbantrees}
\alias{urbantrees}
\title{Sample of tree heights and diameter sizes}
\format{A dataframe with columns containing the \code{'species'},
\code{'height'} (m) and \code{'diameter'} (m).
Each row is a measurement for an individual tree (no tree is represented more than once).
There are 100 measurements per species, across a broad range of height and diameter sizes.}
\usage{
data(urbantrees)
}
\description{
Example data on the size dimensions of five urban tree species
planted along streets in the tropical city of Singapore.
}
\examples{
data(urbantrees)
unique(urbantrees$species) # species in the dataset

table(urbantrees$species) # number of data points (rows) per species
}
\keyword{datasets}
