% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_modelfit.R
\name{sp_modelfit}
\alias{sp_modelfit}
\title{Fit data to a specified linear model for one species}
\usage{
sp_modelfit(data, modelcode, response = "height", predictor = "diameter")
}
\arguments{
\item{data}{Dataframe that contains the variables of interest. Each row is a
measurement for an individual tree.}

\item{modelcode}{Character string of the model code for the selected
allometric equation. Refer to \code{?eqns_info} and \code{data(eqns_info)} for more
information.}

\item{response}{Column name of the response variable. Defaults to \code{height}.}

\item{predictor}{Column name of the predictor variable. Defaults to
\code{diameter}.}
}
\value{
A list of 2 elements:
\enumerate{
\item Resulting model object, named \code{fitted_model}.
\item Table showing information on the resulting model, named \code{fitted_model_info}.
}
\subsection{fitted_model_info}{

\code{fitted_model_info} is a dataframe with the following variables:
\describe{
\item{modelcode}{Model code for the allometric equation used}
\item{a, b, c, d, e}{Parameter estimates}
\item{response_geom_mean}{Geometric mean of the response variable used in
calculation of AICc (only for transformed models)}
\item{correctn_factor}{Bias correction factor to use on model predictions
(only for transformed models)}
\item{predictor_min, predictor_max}{Range of the predictor variable within
the data used to generate the model (i.e. threshold values for model extrapolation)}
\item{residual_SE}{Residual standard error of the model}
\item{mean_SE}{Mean standard error of the model}
\item{adj_R2}{Adjusted \eqn{R^2} of the model}
\item{n}{Sample size (no. of trees used to fit model)}
}
}
}
\description{
Fit data to a specified allometric equation, for one species.
Allometric equations that may be considered as an input to this function
can be found in \code{?eqns_info} and \code{data(eqns_info)}.
}
\examples{
data(urbantrees)
Alb_sam <- urbantrees[urbantrees$species == 'Albizia saman', ]
results <- sp_modelfit(Alb_sam, modelcode = 'quad_w1', # manually specify equation to use
                       response = 'height', predictor = 'diameter')

results$fitted_model

results$fitted_model_info

}
\references{
McPherson E. G., van Doorn N. S. & Peper P. J. (2016) Urban Tree
Database and Allometric Equations. \emph{General Technical Report PSW-GTR-253,
USDA Forest Service}, 86.
}
\seealso{
\code{\link[=sp_modelfit_multi]{sp_modelfit_multi()}} to fit specified models across multiple
species.

\code{\link[=sp_modelselect]{sp_modelselect()}} to select a best-fit model for one species.

\code{\link[=sp_modelselect_multi]{sp_modelselect_multi()}} to select best-fit models across multiple species.

Other single-species model functions: 
\code{\link{sp_modelfit_multi}()},
\code{\link{sp_modelselect_multi}()},
\code{\link{sp_modelselect}()},
\code{\link{sp_predict}()},
\code{\link{sp_simulate}()}
}
\concept{single-species model functions}
