% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-urbantrees.R
\docType{data}
\name{urbantrees}
\alias{urbantrees}
\title{Street tree data in Singapore}
\format{A dataframe. Each row is a measurement for an individual tree.
The following variables are included as columns:
\itemize{
\item \code{species}: Name of tree species
\item \code{height} (m): Visually estimated height
\item \code{diameter} (m): Tree girth measured with diameter tape (at breast-height),
and converted to diameter size
}}
\usage{
data(urbantrees)
}
\description{
Example data on the height and trunk diameter of five tree species
planted along streets in Singapore (sample size = 1585).
}
\examples{
data(urbantrees)
unique(urbantrees$species) # species in the dataset

table(urbantrees$species) # number of data points (rows) per species
}
\keyword{datasets}
